Move scope name related mappings into constants module
